+++ backend/src/main/java/com/cruru/member/service/facade/MemberFacade.java
@@ -1,0 +1,21 @@
+package com.cruru.member.service.facade;
+
+import com.cruru.member.controller.dto.MemberCreateRequest;
+import com.cruru.member.domain.Member;
+import com.cruru.member.service.MemberService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class MemberFacade {
+
+    private final MemberService memberService;
+
+    public long create(MemberCreateRequest request) {
+        Member savedMember = memberService.create(request);
+        return savedMember.getId();
+    }
+}
